requirements.txt
fastapi
uvicorn
websockets
sarvamai
webrtcvad
setuptools
python-dotenv